requirements.txt
# Core agent frameworks
crewai>=0.28.0
pyautogen>=0.4.0
langgraph>=0.1.0
langchain>=0.2.0
langchain-groq>=0.1.0

# LLM
groq>=0.9.0

# Vector store
chromadb>=0.5.0
sentence-transformers>=2.7.0

# Memory
mem0ai>=0.1.0

# API & comms
fastapi>=0.111.0
uvicorn>=0.30.0
python-telegram-bot>=21.0
httpx>=0.27.0

# Data & utils
pydantic>=2.7.0
python-dotenv>=1.0.0
#sqlite3  # stdlib
rich>=13.7.0          # pretty CLI output
loguru>=0.7.0

# Testing
pytest>=8.2.0
pytest-asyncio>=0.23.0
